Extract pure utilities into @sphere/utils workspace package
