+++ apps/sphere/src/hooks/useFeatureProperties.ts
@@ -1,7 +1,7 @@
 import { actions } from "@/store"
 import { useAppDispatch } from "@/store/hooks"
 import { toPropertiesEntries } from "@/lib/properties"
-import { deduplicate } from "@/lib/array"
+import { deduplicate } from "@sphere/utils/array"
 import { useEffect } from "react"
 import type { MapRef } from "react-map-gl/maplibre"
