+++ packages/utils/package.json
@@ -1,0 +1,20 @@
+{
+    "name": "@sphere/utils",
+    "version": "0.0.0",
+    "private": true,
+    "type": "module",
+    "exports": {
+        "./array":             "./src/array.ts",
+        "./color-scheme":      "./src/color-scheme.ts",
+        "./math":              "./src/math.ts",
+        "./once":              "./src/once.ts",
+        "./path":              "./src/path.ts",
+        "./predict-data-type": "./src/predict-data-type.ts",
+        "./stat":              "./src/stat.ts",
+        "./time":              "./src/time.ts"
+    },
+    "scripts": {
+        "typecheck": "tsc --noEmit",
+        "test": "vitest run"
+    }
+}
