+++ packages/utils/src/math.ts
