+++ packages/utils/src/stat.ts
