+++ packages/utils/vitest.config.ts
@@ -1,0 +1,10 @@
+/// <reference types="vitest" />
+import { defineConfig } from "vitest/config"
+
+export default defineConfig({
+    test: {
+        include: ["src/**/*.test.ts", "src/**/*.test.tsx"],
+        globals: true,
+        environment: "happy-dom",
+    },
+})
